+++ src/evm/releases/flow/v1.0/deployments.ts
@@ -117,18 +117,12 @@ export const testnets: Sablier.EVM.Deployment[] = [
     [manifest.FLOW_NFT_DESCRIPTOR]: "0x168ad0b246f604bc70bef87ecde585c3f1d49617",
     [manifest.SABLIER_FLOW]: ["0xd5f78708d83ac2bc8734a8cdf2d112c1bb3b62a2", 18_780_695],
   }),
-  get(chains.blastSepolia.id, {
-    [manifest.FLOW_NFT_DESCRIPTOR]: "0x567a95aa72a23b924f79dfa437d28c38740e144c",
-    [manifest.SABLIER_FLOW]: ["0xa8c864c53e72301c2ab484d013627a5a7084174b", 14_321_690],
-  }),
+
   get(chains.lineaSepolia.id, {
     [manifest.FLOW_NFT_DESCRIPTOR]: "0xcd8871a22640c57ba36984fb57e9c794f5df7f40",
     [manifest.SABLIER_FLOW]: ["0xb0255ed1ee5c01dfe865c1b21bbf56a80f9ae739", 6_731_428],
   }),
-  get(chains.modeTestnet.id, {
-    [manifest.FLOW_NFT_DESCRIPTOR]: "0x1cae76b71913598d7664d16641ccb6037d8ed61a",
-    [manifest.SABLIER_FLOW]: ["0xf5ac60870e1ccc4bfce23cfbb7a796a0d8dbaf47", 22_766_790],
-  }),
+
   get(chains.optimismSepolia.id, {
     [manifest.FLOW_NFT_DESCRIPTOR]: "0x28401987a23ed9b8926b07f3b6855222a70c2128",
     [manifest.SABLIER_FLOW]: ["0x417db0f2bd020fc4d6bccea6b2bb6be0c541862e", 20_763_576],